Pi Data/requirements.txt
ultralytics
python-socketio
websocket
pandas
numpy
pygame

#on the pi
#picamera2
#RPi-GPIO / rpi-lgpio (for pi 5)
